Switch internal pools to use daemon threads
The use of dameon threads will help avoid issues where mariaDb is holding up shutdown of applications.
In addition both pool implementations now default to the same "MariaDbThreadFactory" to reduce code duplication.
